+++ resharper/resharper-unity/src/Rider/Protocol/RdExtensions.cs
@@ -4,7 +4,7 @@ using JetBrains.Collections.Viewable;
 using JetBrains.Lifetimes;
 using JetBrains.Rd.Tasks;
 
-namespace JetBrains.ReSharper.Plugins.Unity.Rider
+namespace JetBrains.ReSharper.Plugins.Unity.Rider.Protocol
 {
     public static class RdExtensions
     {
